+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added visualisations in DS2 (ygraph, bush, map, heatmap, singleValue, tune)
 
 ### Fixed
+- add settings for singleValue in reportsV2
+- fixed reportsV2 heat map visualisation
 - fixed and disabled eslint errors
 - fixed White areas in SingleValue
 - fixed modal-confirm in modal-profile and modal-create